FIx up project structure
