+++ iphone/Classes/MediaModule.h
@@ -9,8 +9,8 @@
 #import "KrollCallback.h"
 #import "MediaPlayer/MediaPlayer.h"
 #import "TiMediaAudioSession.h"
-#import "TiMediaAudioStateConstants.h"
 #import "TiMediaMusicPlayer.h"
+#import "TiMediaTypes.h"
 #import "TiModule.h"
 #import "TiViewProxy.h"
